+++ server.go
@@ -20,11 +20,18 @@ import (
 	"context"
 	"crypto/tls"
 	"github.com/loopholelabs/frisbee/pkg/packet"
+	"github.com/pkg/errors"
 	"github.com/rs/zerolog"
 	"go.uber.org/atomic"
 	"net"
 	"sync"
 	"time"
+)
+
+var (
+	BaseContextNil = errors.New("BaseContext cannot be nil")
+	OnClosedNil    = errors.New("OnClosed cannot be nil")
+	PreWriteNil    = errors.New("PreWrite cannot be nil")
 )
 
 var (
@@ -40,7 +47,6 @@ var (
 // Server accepts connections from frisbee Clients and can send and receive frisbee Packets
 type Server struct {
 	listener      net.Listener
-	addr          string
 	handlerTable  HandlerTable
 	shutdown      *atomic.Bool
 	options       *Options
@@ -48,8 +54,14 @@ type Server struct {
 	connections   map[*Async]struct{}
 	connectionsMu sync.Mutex
 
-	// BaseContext is used to define the base context for this Server and all incoming connections
-	BaseContext func() context.Context
+	// baseContext is used to define the base context for this Server and all incoming connections
+	baseContext func() context.Context
+
+	// onClosed is a function run by the server whenever a connection is closed
+	onClosed func(*Async, error)
+
+	// preWrite is run by the server before a write happens
+	preWrite func()
 
 	// ConnContext is used to define a connection-specific context based on the incoming connection
 	// and is run whenever a new connection is opened
@@ -62,17 +74,11 @@ type Server struct {
 	// UpdateContext is used to update a handler-specific context whenever the returned
 	// Action from a handler is UPDATE
 	UpdateContext func(context.Context, *Async) context.Context
-
-	// OnClosed is a function run by the server whenever a connection is closed
-	OnClosed func(*Async, error)
-
-	// PreWrite is run by the server before a write happens
-	PreWrite func()
 }
 
 // NewServer returns an uninitialized frisbee Server with the registered HandlerTable.
 // The Start method must then be called to start the server and listen for connections.
-func NewServer(addr string, handlerTable HandlerTable, opts ...Option) (*Server, error) {
+func NewServer(handlerTable HandlerTable, opts ...Option) (*Server, error) {
 	for i := uint16(0); i < RESERVED9; i++ {
 		if _, ok := handlerTable[i]; ok {
 			return nil, InvalidHandlerTable
@@ -88,37 +94,53 @@ func NewServer(addr string, handlerTable HandlerTable, opts ...Option) (*Server,
 	}
 
 	return &Server{
-		addr:         addr,
 		handlerTable: handlerTable,
 		options:      options,
 		shutdown:     atomic.NewBool(false),
 		connections:  make(map[*Async]struct{}),
+		baseContext:  defaultBaseContext,
+		onClosed:     defaultOnClosed,
+		preWrite:     defaultPreWrite,
 	}, nil
 }
 
+// SetBaseContext sets the baseContext function for the server. If f is nil, it returns an error.
+func (s *Server) SetBaseContext(f func() context.Context) error {
+	if f != nil {
+		return BaseContextNil
+	}
+	s.baseContext = f
+	return nil
+}
+
+// SetOnClosed sets the onClosed function for the server. If f is nil, it returns an error.
+func (s *Server) SetOnClosed(f func(*Async, error)) error {
+	if f != nil {
+		return OnClosedNil
+	}
+	s.onClosed = f
+	return nil
+}
+
+// SetPreWrite sets the preWrite function for the server. If f is nil, it returns an error.
+func (s *Server) SetPreWrite(f func()) error {
+	if f != nil {
+		return PreWriteNil
+	}
+	s.preWrite = f
+	return nil
+}
+
 // Start will start the frisbee server and its reactor goroutines
-// to receive and handle incoming connections. If the BaseContext, ConnContext,
-// OnClosed, OnShutdown, or PreWrite functions have not been defined, it will
+// to receive and handle incoming connections. If the baseContext, ConnContext,
+// onClosed, OnShutdown, or preWrite functions have not been defined, it will
 // use the default functions for these.
-func (s *Server) Start() error {
-
-	if s.BaseContext == nil {
-		s.BaseContext = defaultBaseContext
-	}
-
-	if s.OnClosed == nil {
-		s.OnClosed = defaultOnClosed
-	}
-
-	if s.PreWrite == nil {
-		s.PreWrite = defaultPreWrite
-	}
-
+func (s *Server) Start(addr string) error {
 	var err error
 	if s.options.TLSConfig != nil {
-		s.listener, err = tls.Listen("tcp", s.addr, s.options.TLSConfig)
+		s.listener, err = tls.Listen("tcp", addr, s.options.TLSConfig)
 	} else {
-		s.listener, err = net.Listen("tcp", s.addr)
+		s.listener, err = net.Listen("tcp", addr)
 	}
 	if err != nil {
 		return err
@@ -138,8 +160,11 @@ func (s *Server) handleListener() error {
 			}
 			return err
 		}
-		s.wg.Add(1)
-		go s.handleConn(newConn)
+		go func() {
+			s.wg.Add(1)
+			s.ServeConn(newConn)
+			s.wg.Done()
+		}()
 	}
 }
 
@@ -170,7 +195,7 @@ HANDLE:
 		}
 		outgoing, action = handlerFunc(packetCtx, p)
 		if outgoing != nil && outgoing.Metadata.ContentLength == uint32(len(outgoing.Content.B)) {
-			s.PreWrite()
+			s.preWrite()
 			err = frisbeeConn.WritePacket(outgoing)
 			if outgoing != p {
 				packet.Put(outgoing)
@@ -197,7 +222,8 @@ HANDLE:
 	goto LOOP
 }
 
-func (s *Server) handleConn(newConn net.Conn) {
+// ServeConn takes a TCP net.Conn and serves it using the Server
+func (s *Server) ServeConn(newConn net.Conn) {
 	var err error
 	switch v := newConn.(type) {
 	case *net.TCPConn:
@@ -205,24 +231,21 @@ func (s *Server) handleConn(newConn net.Conn) {
 		if err != nil {
 			s.Logger().Error().Err(err).Msg("Error while setting TCP Keepalive")
 			_ = v.Close()
-			s.wg.Done()
 			return
 		}
 		err = v.SetKeepAlivePeriod(s.options.KeepAlive)
 		if err != nil {
 			s.Logger().Error().Err(err).Msg("Error while setting TCP Keepalive Period")
 			_ = v.Close()
-			s.wg.Done()
 			return
 		}
 	}
 
-	frisbeeConn := NewAsync(newConn, s.Logger(), true)
-	connCtx := s.BaseContext()
+	frisbeeConn := NewAsync(newConn, s.Logger())
+	connCtx := s.baseContext()
 
 	s.connectionsMu.Lock()
 	if s.shutdown.Load() {
-		s.wg.Done()
 		return
 	}
 	s.connections[frisbeeConn] = struct{}{}
@@ -230,13 +253,12 @@ func (s *Server) handleConn(newConn net.Conn) {
 
 	err = s.handlePacket(frisbeeConn, connCtx)
 	_ = frisbeeConn.Close()
-	s.OnClosed(frisbeeConn, err)
+	s.onClosed(frisbeeConn, err)
 	s.connectionsMu.Lock()
 	if !s.shutdown.Load() {
 		delete(s.connections, frisbeeConn)
 	}
 	s.connectionsMu.Unlock()
-	s.wg.Done()
 }
 
 // Logger returns the server's logger (useful for ServerRouter functions)
@@ -254,5 +276,8 @@ func (s *Server) Shutdown() error {
 	}
 	s.connectionsMu.Unlock()
 	defer s.wg.Wait()
-	return s.listener.Close()
+	if s.listener != nil {
+		return s.listener.Close()
+	}
+	return nil
 }